Add basic testing system, and fix help returning a failure.
If you requested help, that's not a failure - you got what you asked
for.

Testing code has been added now, so that it's possible to introduce
more tests as we go. In theory I should move all the tests that I had
in the original CMunge source in 'justin' repo, but that'll come later.