Update the assembler and preprocessor to use riscos- prefixes for Norcroft.
The objasm and cc tools are no longer called 'objasm' and 'ncc' on non-RISC OS
systems. I've decided to make the entire toolchain prefixed with 'riscos-'
to make it easier to work within other systems. This means that, for the builds
I'm doing, they need to be called properly.

This change introduces a pair of definitions that allow the tools to be called
by the riscos- prefixed names (or to be changed, if that's appropriate).