Updated error messages on command tables to include the command name.
When something went wrong on the command table definition, the error
report would say that there was:

   "Trailing junk after help-text in command-keyword-table"

but not say which command was the one that was problematic. This has
now been added to the error message, to make it easier to find out
what's going wrong.

The error message formatting functions now use the format string
checks built into CLang and GCC to check that there aren't any
missing or mismatched format string parameters.