Fix for the module header not being marked with arch bit.
The module initialisation header isn't marked with the architecture
bit. This meant that the modules shouldn't initialise on RISC OS 64,
(but bugs in Pyromaniac meant that they did) and that that they would
try to initialise on RISC OS 32 but would end up failing because the
code wasn't ARM.

Version number bumped!