Fix for _rmf64_header being doubly defined; version 0.83.
Real fix for the _rm64_header being defined in the zerrors output,
and cleaned up what will be included in those files. In the 64bit
version we just put this in the rodata section, so that we avoid
the problems with it having to be placed in a CODE area (.text).